made more list comprehensions accepted
The brackets are not actually an inherent part of list comprehensions;
there can be multiple comprehensions within one list or nested `for`
clauses. Accordingly, list comprehensions should instead be treated as
list cells rather than top-level expressions.

Examples:
- `[for (i = [1, 2]) i, for (j = [3, 4]) j]` == `[1, 2, 3, 4]`
- `[for (i = [1, 2]) for (j = [3, 4]) i * j]` == `[3, 4, 6, 8]`

The addition of a precedence on `if_clause` is necessary to satisfy the
grammar generator; I must confess that I'm not sure why or whether it
should be right or left.